requirements.txt
# Passport Photo App - Requirements

# ===========================================
# Core Dependencies
# ===========================================

# OpenCV for camera capture and image processing
opencv-python>=4.8.0

# Pillow for image manipulation
Pillow>=10.0.0

# NumPy for numerical operations
numpy>=1.24.0

# ===========================================
# Background Removal (Optional but recommended)
# ===========================================

# rembg for AI-powered background removal
# Install with: pip install rembg
# rembg>=2.0.0

# onnxruntime for rembg (required if using rembg)
# onnxruntime>=1.15.0

# ===========================================
# Web API Dependencies
# ===========================================

# FastAPI for REST API server
fastapi>=0.100.0

# Uvicorn for ASGI server
uvicorn[standard]>=0.23.0

# Pydantic for data validation
pydantic>=2.0.0

# python-multipart for file uploads
python-multipart>=0.0.6

# python-dotenv for environment variables
python-dotenv>=1.0.0
